requirements.txt
Flask==3.1.0
flask-session==0.8.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
msal==1.31.1
requests==2.32.3
python-docx==1.1.2
python-dotenv==1.0.1
lxml==5.3.0
Pillow==11.1.0
psycopg2-binary==2.9.10
anthropic==0.49.0
